backend/requirements.txt
aiohappyeyeballs>=2.6.1
aiohttp>=3.11.16
aiosignal>=1.3.2
asgiref>=3.8.1
attrs>=25.3.0
click>=8.1.8
colorlog>=6.9.0
Django>=5.2
djangorestframework>=3.16.0
frozenlist>=1.5.0
h11>=0.14.0
idna>=3.10
multidict>=6.4.3
pexpect>=4.8
pip>=24.0
propcache>=0.3.1
psutil>=7.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
sqlparse>=0.5.3
uvicorn>=0.34.1
yarl>=1.19.0
